refactor: migrate deck import page
