+++ src/features/import/index.ts
@@ -1,2 +1,3 @@
-export { DeckImportContainer } from "./containers/DeckImportContainer";
+export type { DeckImportPreview, DeckImportResult } from "./components/deckImportTypes";
+export { useDeckImport } from "./hooks/useDeckImport";
 export { useSampleDeckBootstrap } from "./hooks/useSampleDeckBootstrap";
